Deferred email sending, documentation
